+++ packages/common/src/content/_internal/ContentUiSchemaEdit.ts
@@ -88,7 +88,9 @@ export const buildUiSchema = async (
           getThumbnailUiSchemaElement(
             i18nScope,
             options.thumbnail,
-            options.thumbnailUrl
+            options.thumbnailUrl,
+            "content",
+            context.requestOptions
           ),
           // tags
           {