+++ packages/common/src/discussions/_internal/DiscussionUiSchemaEdit.ts
@@ -67,7 +67,9 @@ export const buildUiSchema = async (
           getThumbnailUiSchemaElement(
             i18nScope,
             options.thumbnail,
-            options.thumbnailUrl
+            options.thumbnailUrl,
+            "discussion",
+            context.requestOptions
           ),
         ],
       },