+++ packages/common/test/discussions/_internal/DiscussionUiSchemaEdit.test.ts
@@ -86,6 +86,8 @@ describe("buildUiSchema: discussion edit", () => {
               options: {
                 control: "hub-field-input-image-picker",
                 imgSrc: "https://some-thumbnail-url.com",
+                defaultImgUrl:
+                  "https://www.customUrl/apps/sites/ember-arcgis-opendata-components/assets/images/placeholders/discussion.png",
                 maxWidth: 727,
                 maxHeight: 484,
                 aspectRatio: 1.5,
@@ -278,6 +280,8 @@ describe("buildUiSchema: discussion edit", () => {
               options: {
                 control: "hub-field-input-image-picker",
                 imgSrc: "https://some-thumbnail-url.com",
+                defaultImgUrl:
+                  "https://www.customUrl/apps/sites/ember-arcgis-opendata-components/assets/images/placeholders/discussion.png",
                 maxWidth: 727,
                 maxHeight: 484,
                 aspectRatio: 1.5,